extract WebauthnDescriptionModal into a reusable common component
